+++ example/majority_vote.dart
@@ -4,18 +4,12 @@ import 'package:cellular_automata/src/renderers/ascii_renderer.dart';
 
 // Simple example of cellular_automata.
 // For prettier outputs see the platform specific renderers
-// Run in terminal with command: pub run example/game_of_life.dart
+// Run in terminal with command: pub run example/majority_vote.dart
 void main() {
   final width = 48;
   final height = 22;
 
-  final palette = new Map<GameOfLifeStates, String>.from({
-    GameOfLifeStates.DEAD: ' ',
-    GameOfLifeStates.DEAD_UNDER_POPULATED: '_',
-    GameOfLifeStates.DEAD_OVER_POPULATED: '_',
-    GameOfLifeStates.ALIVE: 'O',
-    GameOfLifeStates.ALIVE_BORN: '*',
-  });
+  final palette = new Map<bool, String>.from({true: '_', false: 'X'});
 
   final AsciiRenderer renderer = new AsciiRenderer();
   final scene = new Scene<String>(
@@ -28,17 +22,17 @@ void main() {
       scene
         ..clearAutomata()
         ..addAutomaton(
-            automaton: new Automaton<GameOfLifeStates, String>(
+            automaton: new Automaton<bool, String>(
           width: width,
           height: height,
-          defaultState: GameOfLifeStates.DEAD,
+          defaultState: false,
           palette: palette,
           wrap: true,
-          rules: new GameOfLife(),
-        )..applyGenerator(new MathematicalGenerator<GameOfLifeStates>(
+          rules: new MajorityVote(),
+        )..applyGenerator(new MathematicalGenerator<bool>(
                 type: MathematicalGenerators.RANDOM,
-                valueTrue: GameOfLifeStates.ALIVE_BORN,
-                valueFalse: GameOfLifeStates.DEAD)));
+                valueTrue: true,
+                valueFalse: false)));
     })
     ..onFullPaint.listen((CellGrid<String> snapshot) {
       renderer
@@ -46,7 +40,7 @@ void main() {
         ..render(snapshot);
 
       print('');
-      print('Rules: Conway\'s Game of Life');
+      print('Rules: Majority Vote');
       print('Generation: ${scene.generationCounter}');
       print('');
     })